Refactor: 탭영역 컴포넌트화
